refactor: вынес TaskManager в отдельный модуль services.py
